Fixed heap corruption issues.
We were shorting ourselves an array cell when we malloc a new buffer.
The side effect was that when we wrote the setinel value to the
buffer, it caused heap corruption due to us technically writing
outside the allocated memory for the buffer.

I'm feeling very much like the amatuer I am.