Implemented syntax error reporting.
There is a lot going on in commit. The first new thing to cover is
the addition of macros for logging information to console. The
second is the rather messy implementation of a DFA for recognizing
the syntax of SUBLEQ assembly. Both of these constitute our syntax
error checking and reporting, which satisfies my goals for this
branch.